+++ my-app/src/App.js
@@ -12,7 +12,7 @@ import "./App.css";
 import { Cart } from "./Pages/Cart/cart";
 import { Nav } from "./components/nav/nav";
 import { ProductPage } from "./Pages/Productpage/productpage";
-import { Wishlist } from "./components/wishlist";
+import { Wishlist } from "./Pages/Wishlist/wishlist";
 import { Homepage } from "./Pages/Homepage/Homepage";
 import { Login } from "./Pages/Login/Login";
 import { Register } from "./Pages/Register/Register";
